+++ R/get_pmids_from_one_search.R
@@ -36,6 +36,16 @@
 get_pmids_from_one_search <- function (query, api_key) {
 
     out <- tryCatch({
+
+        ## Check that API key is well-formed
+        api_key <- stringr::str_trim(api_key)
+        assertthat::assert_that(
+                        grepl(
+                            "^[0-9a-f]{36}$",
+                            as.character(api_key)
+                        ),
+                        msg="Pubmed API key is not well-formed"
+                    )
         
         pubmed_search <- list(
             api_key = api_key,