+++ NEWS.md
@@ -1,3 +1,14 @@
+# pubmedtk 1.0.2
+
+* Added optional progress feedback during download for
+  `get_pmids_from_searches()` and `get_metadata_from_pmids()`
+* Added `$pubdate` and `$epubdate` download to
+  `get_metadata_from_one_pmid()` and corresponding columns to
+  `get_metadata_from_pmids()`
+* Improved feedback for non-valid input
+
+---
+
 # pubmedtk 1.0.1
 
 * Fix bug with uncommented instructions in code